requirements.txt
click>=8.1.7
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.3
